Exo météo avec import de classes et de variable.
