+++ js/js/assets/CustomDom.js
@@ -1,0 +1,20 @@
+export default class CustomDom {
+   /**
+   * Crée un élément du dom, lui ajoute du texte, le place comme dernier
+   * enfant de parent et ajout un attribut en utilisant le paramètre attribute
+   * @param {String} markup_name 
+   * @param {String} text 
+   * @param {domElement} parent 
+   * @param {Object} attribute  (doit comprendre les propriétés name et value)
+   * @returns domElement
+   */
+    createMarkup(markup_name, text, parent, attribute) {
+      const markup = document.createElement(markup_name);
+      markup.textContent = text;
+      parent.appendChild(markup);
+      if (attribute && attribute.hasOwnProperty("name")) {
+        markup.setAttribute(attribute.name, attribute.value);
+      }
+      return markup;
+    }
+}
